+++ CHANGELOG.md
@@ -31,3 +31,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed parameter 'Name' to 'Names' in 'WindowsFeatures' Resource according to coding convention.
 - Made the Office Online Server resources actually work and redesigned them.
   - Added 'OfficeOnlineServerMachineConfig' configuration.
+- WindowsFeatures configuration does not longer install all sub features. If needed, use prefix '*'.